Move node description, aux fields, and slot helpers to cluster.c
Move protocol-agnostic functions from cluster_legacy.c to cluster.c:
- CLUSTER SHARDS: clusterCommandShards, addNodeDetailsToShardReply
- CLUSTER NODES: clusterGenNodeDescription, clusterGenNodesDescription,
  representClusterNodeFlags, representSlotInfo
- Slot info helpers: clusterGenNodesSlotsInfo, clusterFreeNodesSlotsInfo
- Aux field handlers and all getter/setter/present functions

Add vtable callbacks for protocol-specific behavior:
- getNodePingPongEpoch: per-node ping/pong/epoch for CLUSTER NODES output
- setNodePingPongEpoch/setNodeFailed: set per-node fields during load
- appendVarsLine/parseVarsLine/postLoad: nodes.conf vars line handling
- initNodeData: allocate protocol-specific node data
- onSlotReset: slot assignment bookkeeping

Refactor createClusterNode to use initNodeData vtable callback.
Refactor clusterAddSlot/clusterDelSlot to use onSlotReset callback.
